Add some basic tests.
These tests are meant to cover some of the things which might go wrong
with the nose_xunitmp plugin. They are not intended to cover the
functionality in the xunit plugin that it subclasses.

Due to how PluginTester works, the tests have to use the multiprocessing
plugin's `_instantiate_plugins` in order to ensure that the plugin is
loaded in the subprocesses. In normal usage of nose_xunitmp, this is not
required and the nose_xunitmp will be loaded (re-instantiated) in each
subprocess.