Make the tests run on Windows too.
Unfortunately, nose.plugins.multiprocessing._instantiate_plugins does
not work on Windows, so we need to use another hack to trick
PluginTester into working with multiprocessing.

This change allows the tests to run on Windows and I've confirmed
that the change in c74210cce9 allows nose_xunitmp to run on Windows.